Add files via upload
Mini Project 1
Statement:-
The task is to create an “Online Library Management System.” 
For this, you have to create a library class that includes the 
following methods:
Displaybook(): To display the available books
Lendbook(): To lend a book to a user
Addbook(): To add a book to the library
Returnbook(): To return the book to the library.
As you have created a library class, now you will create an 
object and pass the following parameters in the constructor.
HarryLibrary=Library(listofbooks, library_name)
After that, create a main function and run an infinite while loop 
that asks the users for their input whether they want to display, 
lend, add or return a book.
Optional:-
Maintain a dictionary for the users who own a book. Dictionary 
should take book name as a key and name of the person as a 
value. Whenever you lend a book to a user, you should maintain 
a dictionary